Make sign message testable
